+++ src/app/admin/courses/QuickEditCourseForm.tsx
@@ -2,20 +2,22 @@
 
 import * as Yup from "yup";
 import { useMutation, useQuery } from "convex/react";
-import { api } from "../../../convex/_generated/api";
+import { api } from "../../../../convex/_generated/api";
 import {
+  AdminFieldtype,
   AdminQuickForm,
   AdminFormConfig,
-} from "../../app/admin/AdminQuickForm";
+} from "../AdminQuickForm";
 import { useToast } from "@/lib/hooks/useToast";
-import { Id } from "../../../convex/_generated/dataModel";
+import { Id } from "../../../../convex/_generated/dataModel";
 import { Pencil } from "lucide-react";
 import { Button, Tooltip, TooltipContent, TooltipTrigger } from "@/lib/ui";
 
 // Define the fields that can be edited
-export interface Assessment {
+export interface Course {
   title: string;
   description: string;
+  isPublished: boolean;
 }
 
 // Define the validation schema
@@ -26,27 +28,26 @@ const validationSchema = Yup.object().shape({
   description: Yup.string()
     .max(50, "Description must be less than 50 characters")
     .required("Description is required"),
+  isPublished: Yup.boolean().optional(),
 });
 
 // Set toast messages for success and error
-const successMessage = "Assessment saved.";
-const errorMessage = "Something went wrong trying to update assessment.";
+const successMessage = "Course saved.";
+const errorMessage = "Something went wrong trying to edit course.";
 
 // Set the form title
-const formTitle = "Edit assessment";
+const formTitle = "Edit course";
 
-interface EditAssessmentFormProps {
-  assessmentId: Id<"assessments">;
+interface EditCourseFormProps {
+  courseId: Id<"courses">;
 }
 
-export const QuickEditAssessmentForm = ({
-  assessmentId,
-}: EditAssessmentFormProps) => {
-  // Fetch the assessment to edit
-  const assessment = useQuery(api.assessments.findById, { id: assessmentId });
+export const QuickEditCourseForm = ({ courseId }: EditCourseFormProps) => {
+  // Fetch the course to edit
+  const course = useQuery(api.courses.findById, { id: courseId });
 
   // Define the mutation
-  const editAssessment = useMutation(api.assessments.update);
+  const editCourse = useMutation(api.courses.update);
 
   const { toast } = useToast();
 
@@ -54,11 +55,17 @@ export const QuickEditAssessmentForm = ({
   const fields = [
     { name: "title", label: "Title", initialValue: "" },
     { name: "description", label: "Description", initialValue: "" },
+    {
+      name: "isPublished",
+      label: "Published?",
+      fieldtype: "switch" as AdminFieldtype,
+      initialValue: false,
+    },
   ];
 
-  async function onSubmit(values: Assessment) {
-    if (!assessment) return;
-    const result = await editAssessment({ id: assessmentId, ...values });
+  async function onSubmit(values: Course) {
+    if (!course) return;
+    const result = await editCourse({ id: courseId, ...values });
 
     if (result) {
       toast({
@@ -76,11 +83,12 @@ export const QuickEditAssessmentForm = ({
 
   // Set initial values
   const initialValues = {
-    title: assessment?.title ?? "",
-    description: assessment?.description ?? "",
+    title: course?.title ?? "",
+    description: course?.description ?? "",
+    isPublished: course?.isPublished ?? false,
   };
 
-  const formConfig: AdminFormConfig<Assessment> = {
+  const courseFormConfig: AdminFormConfig<Course> = {
     validationSchema,
     initialValues,
     fields,
@@ -88,8 +96,8 @@ export const QuickEditAssessmentForm = ({
   };
 
   return (
-    <AdminQuickForm<Assessment>
-      config={formConfig}
+    <AdminQuickForm<Course>
+      config={courseFormConfig}
       formTitle={formTitle}
       renderTrigger={
         <Button variant="ghost" size="sm">
@@ -97,7 +105,7 @@ export const QuickEditAssessmentForm = ({
             <TooltipTrigger>
               <Pencil className="h-4 w-4" />
             </TooltipTrigger>
-            <TooltipContent>Edit assessment settings</TooltipContent>
+            <TooltipContent>Edit course settings</TooltipContent>
           </Tooltip>
         </Button>
       }
