+++ src/app/admin/courses/[id]/page.tsx
@@ -6,10 +6,10 @@ import { Id } from "../../../../../convex/_generated/dataModel";
 import { PageContent, PageHeader } from "@/lib/layout";
 import { FlexRow, Text } from "@/lib/ui";
 import { useEffect, useState } from "react";
-import { CourseModulesNav } from "@/lib/Courses/CourseModulesNav";
-import { EditModuleSectionForm } from "@/lib/Modules/EditModuleSectionForm";
-import { QuickEditCourseForm } from "@/lib/Courses/QuickEditCourseForm";
-import { QuickEditModuleForm } from "@/lib/Modules/QuickEditModuleForm";
+import { CourseModulesNav } from "./CourseModulesNav";
+import { EditModuleSectionForm } from "../../modules/[id]/EditModuleSectionForm";
+import { QuickEditCourseForm } from "../QuickEditCourseForm";
+import { QuickEditModuleForm } from "../../modules/QuickEditModuleForm";
 
 interface AdminCoursePageProps {
   params: { id: string };
