+++ src/app/admin/modules/[id]/ModuleSectionsNav.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { useMutation, useQuery } from "convex/react";
-import { api } from "../../../convex/_generated/api";
-import { Id } from "../../../convex/_generated/dataModel";
+import { api } from "../../../../../convex/_generated/api";
+import { Id } from "../../../../../convex/_generated/dataModel";
 import {
   Button,
   DropdownMenu,
