+++ src/app/admin/modules/page.tsx
@@ -1,8 +1,8 @@
 "use client";
 
 import { PageContent, PageHeader } from "@/lib/layout";
-import { CreateModuleForm } from "@/lib/Modules/CreateModuleForm";
-import { ModulesTable } from "@/lib/Modules/ModulesTable";
+import { CreateModuleForm } from "./CreateModuleForm";
+import { ModulesTable } from "./ModulesTable";
 import { Text } from "@/lib/ui";
 
 export default function AdminModulesPage() {
