+++ src/lib/layout/index.ts
@@ -3,4 +3,4 @@ export * from "./Main";
 export * from "./Masthead";
 export * from "./PageContent";
 export * from "./PageHeader";
-export * from "../Authorization/PrivatePage";
+export * from "../authorization/PrivatePage";
