--- src/lib/modules/CreateModuleSectionButton.tsx
@@ -1,3 +1,0 @@
-
-
-
